requirements.txt
RPi.GPIO==0.7.0
Django==2.2.5
smbus2==0.2.3
smbus==1.1.post2
